SVCPLAN-1825 telegraf group by account
Pulls in upstream changes from:
https://git.ncsa.illinois.edu/ici-monitoring/ici-developed-checks

Specifically commits:
https://git.ncsa.illinois.edu/ici-monitoring/ici-developed-checks/-/commit/364e20d7646ce7e939088c59701c713741fbe6e6
https://git.ncsa.illinois.edu/ici-monitoring/ici-developed-checks/-/commit/d5e59c48f2a5d385952ee06881419bc097ae9556

Main changes are:
* A lot of slurm stats were grouped by user, group, and partition. Instead
this is changed to group by "account" since that’s really how we do
allocations on systems, not via their group.  All these groups show up
as 202 since that is everyone’s default gid

* Average and Max job wait time was based on subtracting job submit time
from when it ran, not job eligible time so it looked worse than it is;
since it included time job waited due to job chains/queue limits/etc.
Script has been updated to account for this