SVCPLAN-3768: add passwd_check
Exec a check of identity services (e.g., sssd), which writes a
receipt file if successful. This check queries a configurable group.
This receipt file is a dependency for the slurmd and slurmctld systemd
unit files, to ensure they do not start if identity services are not
(yet) working. (E.g., we don't want slurmctld to flush reservations if
it doesn't recognize users or groups.)